Little ECS that prepares a list to render any amount of meshes
Also, texture system is almost ready, just needs to get loaded properly

+++ src/engine/scene/loader.h
@@ -2,9 +2,8 @@
 
 #include "../renderer/renderer.h"
 
-
 Mesh load_obj_mesh(Renderer* renderer, char* file_path);
-Mesh* load_glft_meshes(Renderer* renderer, char* file_path, int* out_n);
+Mesh* load_glft_meshes(Renderer* renderer, char* file_path, uint8_t* out_n);
 
 void meshes_destroy(Mesh* meshes, int n, VmaAllocator allocator);
